refactor: delete repeatable part of components
