+++ src/components/pokemon/_pokemonList.scss
