+++ src/components/pokemon/pokemonCardAdditional/PokemonCardAdditional.tsx
@@ -1,0 +1,67 @@
+import { PuffLoader } from 'react-spinners';
+import { IPokemon } from '../../../types';
+import PokemonBaseInfo from '../pokemonBaseInfo/PokemonBaseInfo';
+
+const PokemonCardAdditional = ({
+  pokemonFullInfo,
+  searchParams,
+  isLoading,
+  setSearchParams,
+}: {
+  pokemonFullInfo: IPokemon | undefined;
+  searchParams: URLSearchParams;
+  isLoading: boolean;
+  setSearchParams: CallableFunction;
+}) => {
+  const wrapperClass = searchParams.get('details')
+    ? 'pokemon-additional-info pokemon-additional-info_active'
+    : 'pokemon-additional-info';
+
+  const setUrlParams = () => {
+    if (searchParams.get('search')) {
+      setSearchParams({
+        search: searchParams.get('search'),
+      });
+    } else if (searchParams.get('page')) {
+      setSearchParams({ page: searchParams.get('page') });
+    }
+  };
+
+  return (
+    <div className={wrapperClass}>
+      <div className={'pokemon-additional__wrapper'}>
+        {isLoading ? (
+          <PuffLoader color="#ad5905" size={150} className="spinner" />
+        ) : (
+          <>
+            <span
+              className="pokemon__close-menu"
+              onClick={() => setUrlParams()}
+            >
+              Close
+            </span>
+
+            <li className="pokemon__item">
+              <div className="pokemon__info">
+                <PokemonBaseInfo pokemonFullInfo={pokemonFullInfo} />
+                <ul className="pokemon-stats__list">
+                  <span className="pokemon__info-header">Stats:</span>
+                  {pokemonFullInfo?.stats.map(({ base_stat, stat }, index) => (
+                    <li key={index}>
+                      {index + 1 < pokemonFullInfo?.stats.length
+                        ? stat.name + ': ' + base_stat + ','
+                        : stat.name + ': ' + base_stat}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            </li>
+          </>
+        )}
+      </div>
+      <div className="pokemon__overlay" onClick={() => setUrlParams()}></div>
+    </div>
+  );
+};
+
+export default PokemonCardAdditional;
