+++ src/index.scss
@@ -10,7 +10,7 @@
 @import './components/notFound/notFound';
 @import './components/app/app';
 @import './components/errorBoundary/errorBoundary';
-@import './components/itemList/pokemonList';
+@import './components/pokemon/pokemonList';
 @import './components/search/search';
 @import './components/itemPerPage/itemPerPage';
